running project

+++ src/Shop/Product/View/ProductPriceInterface.php
@@ -4,11 +4,11 @@ namespace Mateusz\Mercetree\Shop\Product\View;
 
 interface ProductPriceInterface
 {
-    public function getBasePriceNet() : int;
-    
+    public function getBasePriceNet() : float;
+
     public function getTaxRate() : int;
-    
+
     public function getCurrencySymbol() : string;
-    
+
     public function getQuantity() : int;
 }
